+++ apps/desktop/src/store/workspace/actions/workspace-fs-note-actions.test.ts
@@ -3,21 +3,6 @@ import { createWorkspaceActionTestContext } from "./workspace-action-test-helper
 import { createWorkspaceFsNoteActions } from "./workspace-fs-note-actions"
 
 describe("workspace-fs-note-actions", () => {
-	it("registerLocalMutation registers targets when workspace is active", () => {
-		const { context, originJournal, setState } =
-			createWorkspaceActionTestContext()
-		const actions = createWorkspaceFsNoteActions(context)
-		setState({ workspacePath: "/ws" })
-
-		actions.registerLocalMutation([{ path: "/ws/a.md", scope: "exact" }])
-
-		expect(originJournal.register).toHaveBeenCalledWith({
-			workspacePath: "/ws",
-			targets: [{ path: "/ws/a.md", scope: "exact" }],
-			ttlMs: undefined,
-		})
-	})
-
 	it("saveNoteContent records local mutation as exact path", async () => {
 		const { context, getState, setState } = createWorkspaceActionTestContext()
 		const actions = createWorkspaceFsNoteActions(context)
